fix(nvim): Update script to symlink vi to nvim only if not already done

+++ run_once_after_B-symlink_vi_neovim.sh
@@ -1,4 +1,13 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-sudo ln -s "$(which nvim)" /usr/local/bin/vi
+TARGET_VI=/usr/local/bin/vi
+
+if ! [[ "$(readlink "${TARGET_VI}")" =~ .*/nvim ]]; then
+  echo "Removing ${TARGET_VI}..."
+  sudo rm -v "${TARGET_VI}"
+fi
+
+if ! [ -e "${TARGET_VI}" ]; then
+  sudo ln -s "$(which nvim)" /usr/local/bin/vi
+fi
